Kotlin's version has been raised. Added permission for the geolocation library. The user model has been fixed - an unused field has been removed, and a field for the city has been added.

+++ lib/features/auth/data/repositories/auth_repository_impl.dart
@@ -31,9 +31,14 @@ class AuthRepositoryImpl extends AuthRepository {
 
   @override
   Future<void> logout() async {
-    final auth = await _localDataSource.getAuthModel();
-    _localDataSource.deleteAuthModel();
-    _remoteDataSource.logout(auth: auth.toEntity());
+    AuthModel? auth;
+    try {
+      auth = await _localDataSource.getAuthModel();
+    } catch (e) {
+      if (auth == null) return;
+      _localDataSource.deleteAuthModel();
+      _remoteDataSource.logout(auth: auth.toEntity());
+    }
   }
 
   @override
